Corregidos controladores de una línea estratégica para que se tengan en cuenta los ids de planes no existentes y se lance el respectivo error.

+++ icasus/app_code/linea_borrar.php
@@ -18,23 +18,32 @@ if (filter_has_var(INPUT_GET, 'id_linea') && filter_has_var(INPUT_GET, 'id_entid
     $id_linea = filter_input(INPUT_GET, 'id_linea', FILTER_SANITIZE_NUMBER_INT);
     $id_entidad = filter_input(INPUT_GET, 'id_entidad', FILTER_SANITIZE_NUMBER_INT);
 
+    //Obtenemos los datos de la línea
     $linea = new Linea();
-    $linea->load("id = $id_linea");
-    $id_plan = $linea->id_plan;
-    $plan = new Plan();
-    $plan->load("id=$id_plan");
-    $objest = new ObjetivoEstrategico();
-    $objest_linea = $objest->Find("id_linea=$id_linea");
-    if (!$objest_linea && $control)
+    if ($linea->load("id = $id_linea"))
     {
-        $exito = MSG_LINEA_BORRADA . ' ' . $linea->indice . '. ' . $linea->nombre;
-        $linea->delete();
-        header("Location: index.php?page=plan_mostrar&id_plan=$id_plan&id_entidad=$id_entidad&exito=$exito");
+        $id_plan = $linea->id_plan;
+        $plan = new Plan();
+        $plan->load("id=$id_plan");
+        $objest = new ObjetivoEstrategico();
+        $objest_linea = $objest->Find("id_linea=$id_linea");
+        if (!$objest_linea && $control)
+        {
+            $exito = MSG_LINEA_BORRADA . ' ' . $linea->indice . '. ' . $linea->nombre;
+            $linea->delete();
+            header("Location: index.php?page=plan_mostrar&id_plan=$id_plan&id_entidad=$id_entidad&exito=$exito");
+        }
+        else
+        {
+            $aviso = MSG_LINEA_BORRAR;
+            header("Location: index.php?page=linea_mostrar&id_entidad=$id_entidad&id_linea=$id_linea&aviso=$aviso");
+        }
     }
     else
     {
-        $aviso = MSG_LINEA_BORRAR;
-        header("Location: index.php?page=linea_mostrar&id_entidad=$id_entidad&id_linea=$id_linea&aviso=$aviso");
+
+        $error = ERR_LINEA_MOSTRAR;
+        header("location:index.php?page=error&error=$error");
     }
 }
 else
